Fix for contrib module handling. (#226)
## DESCRIPTION OF CHANGES:.
All the module files contained in this repository have been updated for the change on Jet/Hera to remove the contrib module, and instead use the `module use -a` method for locating necessary contrib package modulefiles. This will be the ONLY way to use these modules after June 11, 2020, and is already in place to work now. As such, this is a **time sensitive** PR.

## TESTS CONDUCTED:.
Loaded a few different modules on front end node. This included modules that need to be loaded, and those "modules" that are actually shell scripts and must be sourced. Both methods worked as expected.

End-to-end tests completed successfully on Hera: 
new_JPgrid
regional_002
regional_003
regional_004
regional_005
regional_006
regional_007
regional_008
regional_015