Updating hashes of regional_workflow, ufs-weather-model, and UFS_UTILS to top of release branches (#97)
## DESCRIPTION OF CHANGES: 
Update hashes; this includes several new capabilities and bugfixes. This constitutes most of the final effort needed for the code freeze, including the the full stand-alone linux and MacOS capabilities, and bug fixes to both the weather model and UFS_UTILS.

## TESTS CONDUCTED: 
Built and ran the release branch end-to-end tests on Hera and Jet. Also built and ran the new stand-alone (non-rocoto) capability on MacOS (gnu 9.3.0) and Cheyenne (Intel 19.1.1). All passed with flying colors!