Enable AWS Parallel Works platform and Add Comprehensive End-To-End Tests (#333)
* Normalize Parallel Works cluster platform value

Set the value of platform to 'noaacloud' when SRW_PLATFORM matches a
Parallel Works cluster name.

* Enable the AWS Parallel Works platform

* Move agent declaration to stages

This change allows the platform filter to work correctly, otherwise, the
Parallel Works clusters would block indefinitely waiting to execute the
matrix on a agent/node that was not started.

* Ensure PROJ_LIB is set on Parallel Works platforms

* Fix final exit status of srw_test script

Some platforms do not recognize quoted variables within an arithmetic
expression. This change removes the quotes.

* Add comprehensive end-to-end tests option

* Add a parameter to the Jenkins pipeline that allows the comprehensive
set of workflow and end-to-end tests to be executed during the test
stage.
* Add logic to the Jenkins pipeline that checks for a specific Pull
Request label, then overrides the comprehensive end-to-end test
parameter's value if set.

* Clean up the workspace after a we2e test run

The experiments directory uses a lot of disk space. Removing it after
the end-to-end tests complete will allow us to keep the workspaces
longer. However, the test logs should be preserved. This change creates
a tarball containing the test logs in the workspace, which is archived,
then removes the experiments directory.

* Disable concurrent builds for branches and PRs

Prevent Jenkins from executing multiple pipelines at the same time for a
given branch or change request.

* Disable branch indexing triggers for pipeline

* Update we2e fundamental tests for srw_test script

* Log and archive the output of the srw build

* Update we2e comprehensive tests in srw_test script

* Update we2e fundamental/comprehensive tests

* Update we2e comprehensive tests

* Remove invalid tests from comprehensive list

* Update Parallel Works cluster names

* Remove regional_workflow from we2e_test_dir path