Fix setting of SRW_WE2E_COMPREHENSIVE_TESTS var
The SRW_WE2E_COMPREHENSIVE_TESTS variable was incorrectly set by a sh
block that proceeded the call to srw_test.sh. This issue was hidden
because Jenkins sets the parameters to environment variables, which are
accessible from sh blocks. However, the parameters do not seem to be set
as environment variables the first time a multi-branch pipeline is
initialized. This resulted in an unbound variable error when the
SRW_WE2E_COMPREHENSIVE_TESTS variable was accessed by the srw_test.sh
script. This update sets SRW_WE2E_COMPREHENSIVE_TESTS in the same sh
block as the srw_test.sh script call, while ensuring the WORKSPACE
variable isn't evaluated until the sh block.