[develop] Integrate 'online-cmaq' into 'develop' branch: Part 2, build script and module files (#549)
Currently, the Online-CMAQ for air quality modeling is separately managed in the 'online-cmaq' branch of the authoritative UFS SRW App repository. The main target of the current 'online-cmaq' branch is the implementation of AQM ver.7 and delivery to NCO in late March. Therefore, it has been developed and tested on WCOSS2 and Hera. The Online-CMAQ checks out the same ufs weather model as the 'develop' branch. However, since it couples another component 'AQM', the compile process is different from that of the 'develop' branch.

To compile Online-CMAQ:

For Hera - ./devbuild.sh -p=hera -a=ATMAQ
For WCOSS2 - ./devbuild.sh -p=wcoss2 -a=ATMAQ

---------

Co-authored-by: chan-hoo <chan-hoo.jeon@clogin04.cactus.wcoss2.ncep.noaa.gov>
Co-authored-by: chan-hoo <chan-hoo.jeon@clogin03.cactus.wcoss2.ncep.noaa.gov>
Co-authored-by: chan-hoo jeon <chan-hoo.jeon@dlogin02.dogwood.wcoss2.ncep.noaa.gov>
Co-authored-by: Michael Lueken <michael.lueken@noaa.gov>