[develop] Add correct logic for 3-digit forecast hours in post (#607)
Adding correct logic for 3-digit forecast hours in UPP. The filenames coming out of UPP are unfortunately hard-coded in that all forecast hours below 100 will have a 2-digit forecast hour in the filename, and beyond that will switch to a 3-digit forecast hour. The current workflow does not attempt to resolve this difference, resulting in UPP tasks failing after (despite the executable completing successfully). This change adds the correct logic to allow UPP tasks to succeed beyond 99 hours.