[develop] Add GDAS and GEFS datasets to the workflow (#526)
This PR adds the ensemble (GDAS and GEFS) datasets as options for the SRW App and is part of a bigger effort to merge in RRFS capabilities into the develop branch.

Currently, the workflow ensemble tasks all use the same initial conditions. Because of the addition of the ensemble datasets, each ensemble task can now have its own initial conditions with the global ensemble number mapping to the SRW App ensemble number (i.e. GDAS mem002 corresponds to SRW App mem002).

The appropriate logic was added to these files so that the GEFS and GDAS datasets can run with the workflow without breaking existing functionality. A number of changes were made in the retrieve_data.py script to handle the --members argument and GEFS filename list. A GEFS function was added to the exregional_get_extrn_mdl_files.sh file to merge the GEFS files and update the summary extrn_mdl_var_defns.sh script.

---------

Co-authored-by: Edward Snyder <Edward.Snyder@noaa.com>
Co-authored-by: Michael Lueken <michael.lueken@noaa.gov>