[develop] Update default timesteps (DT_ATMOS) to less wasteful values (#624)
Based on discussion in last week's code management meeting, it was agreed that the default values for the atmospheric timestep DT_ATMOS are too low, and for most cases are very wasteful of core hours. This change increases the default timestep for the following pre-defined domains:

* RRFS_CONUS_25km, RRFS_CONUScompact_25km: 40 --> 180
* RRFS_CONUS_13km, RRFS_CONUScompact_13km: 45 --> 75
* RRFS_NA_13km: 50 --> 75
* RRFS_CONUScompact_3km, RRFS_SUBCONUS_3km, SUBCONUS_Ind_3km: 40 --> 36
* CONUS_3km_GFDLgrid: 18 --> 36

There is a known limitation for certain CCPP suites designed for high-resolution cases (per @JeffBeck-NOAA, these are FV3_RRFS_v1beta, FV3_WoFS_v0, and FV3_HRRR); these suites may see failures if DT_ATMOS is too high. For those cases, the default DT_ATMOS is set to 40, however users can still over-write this value in their config file.

NOTE: This change will affect results for users who are using the affected domains and have not manually specified DT_ATMOS. This should be reflected in the release notes for the next release.