[develop] Update ufs-weather-model and UPP hash to correct post control file issue (#699)
The March 24, 2023 update to the ufs-weather-model (36d6e16), updated UPP and corrected issues associated with the post control file. The ufs-weather-model hash has been updated to 36d6e16 and the UPP hash has been updated to 22cfb88.

The changes required to allow the previous versions of ufs-weather-model and UPP to run the run_post tasks were removed.