[develop] Add ptype obs (#757)
As part of the (soon-to-end 😢) DTC UFS SRW project, one of our final deliverables was to add precipitation-type verification to MET. This PR modifies the Metplus config files to output the relevant ptype statistics for PB2NC and PointStat.

The precipitation-type variables used are described in the METplus precipitation documentation page. This method uses the reported "PRWE" (present weather) field compared to categorical precipitation variables output from the weather model.