[develop] Move all unittest tests to a common area. (#728)
Moves all unittest tests to the tests directory, and lints them. They currently should pass at 10/10. This makes running them easier so that users may be encouraged and enabled to do that on their own platform.

Adds the HPSS functional tests to the Jenkinsfile, so that the full functionality is tested regularly. These tests run on the front end so shouldn't eat into allocation. The tests clean up after themselves, so shouldn't be a data issue in the long-run.

Adds linting for the tests directory to the GitHub Actions workflow to ensure they stay linted.

---------

Co-authored-by: Michael Lueken <63728921+MichaelLueken@users.noreply.github.com>