[develop] Add USE_USER_STAGED_EXTRN_FILES: true to WE2E configurations that are missing this variable (#780)
During the extended HPSS downtime, it was noted that several tests were failing because the necessary staged data was not being pulled from the expected staged data locations. Upon careful inspection of the WE2E test configuration files for the tests that were failing showed that these tests didn't have the USE_USER_STAGED_EXTRN_FILES: true variable set. Since the rest of the WE2E test configurations have this variable set, I have gone through and set this variable for the tests that it isn't set in.