[develop] Fix retrieve data. (#810)
Fix a whitespace issue that leads to pre-mature exit when looking through the possible archive internal directories.

Also addresses an issue uncovered when a user has a pre-existing PYTHONPATH set. We should be prepending the path instead of appending it so that we override any other clones a user might have in their PYTHONPATH.

This fixes an issue with the nco_grid_RRFS_CONUS_25km_ics_FV3GFS_lbcs_FV3GFS_timeoffset_suite_GFS_v16 test.