[develop] Introducing forecast metrics into CI testing processes (#746)
The main goal of this PR is to provide a demo for integration of forecast metrics into testing processes to fulfill the requirements of NOAA-EPIC deliverable ECC-321 (TO3 PWS 5.6.5. Develop a plan to integrate NOAA baseline operational metric into testing processes.).

This PR contains a script called srw_metric_example.sh. It conducts a forecast run for the Indianapolis Severe Weather case, and results are compared with the reference (the one derived from the v.2.1.0 release) using the Skill Score Index. This work has the potential to be integrated into the Jenkins CI/CD pipeline.

In this example, the skill score index is calculated as a weighted average of 16 skill scores of RMSE statistics for wind speed, dew point temperature, temperature, and pressure at the lowest level in the atmosphere over a 48-hour lead time. A value greater than 1.0 indicates that the forecast model outperforms the reference, while a value less than 1.0 indicates that the reference outperforms the forecast.