[develop] Change Jenkins test to build with "all" target. (#832)
Extends the Jenkins build so that RRFS components may be added to coverage tests.

Also adds the process_obs test to the Jet coverage suite. Jet is chosen here since most of manual testing and development is done on Hera, it should help diversify the platforms GSL is most interested in.