[develop] Change the configuration variable NET to NET_default. (#868)
The renaming of the configuration variable NET to NET_default by PR #828 broke the WE2E test MET_ensemble_verification_only_vx_time_lag (see Issue #851). This PR fixes the bug and adds that WE2E test to the list of coverage tests on Hera (currently that test is only in the comprehensive tests list).