[develop] Update wrapper scripts (#883)
The wrapper scripts are a set of scripts that run the tasks from a SRW experiment. They are used in a situation where rocoto is not installed or available to use. These set of wrapper scripts have become outdated and don't work in its current state. This PR was created to address this issue.

---------

Co-authored-by: Edward Snyder <Edward.Snyder@noaa.com>
Co-authored-by: Eddie Snyder <esnyder@derecho2.hsn.de.hpc.ucar.edu>