[develop] Add integration test job (#1042)
This update adds a test job to the workflow. It was originally written with pytest but because of some file naming issues, the python package unittest was used instead. The test checks for the existence of netcdf files from the weather model.

The necessary scripts were added or modified to incorporate the integration job into the workflow. A wrapper script was also added.

---------

Co-authored-by: Parallel Works app-run user <Edward.Snyder@mgmt-edwardsnyder-aws-00107.pw-noaa-us-east-1.pw.local>
Co-authored-by: michael.lueken <Michael.Lueken@noaa.gov>